Extract duplicate code
