+++ src/main/java/com/rarchives/ripme/ripper/AbstractRipper.java
@@ -11,9 +11,11 @@ import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 
@@ -63,6 +65,8 @@ public abstract class AbstractRipper
     public abstract String getHost();
 
     public abstract String getGID(URL url) throws MalformedURLException, URISyntaxException;
+
+    protected abstract boolean allowDuplicates();
 
     public boolean hasASAPRipping() {
         return false;
@@ -267,7 +271,66 @@ public abstract class AbstractRipper
         return addURLToDownload(tug, ripUrlId, directory, null, referrer, cookies, getFileExtFromMIME);
     }
 
-    public abstract boolean addURLToDownload(TokenedUrlGetter tug, RipUrlId ripUrlId, Path directory, String filename, String referrer, Map<String,String> cookies, Boolean getFileExtFromMIME);
+    /**
+     * Queues multiple URLs of single images to download from a single Album URL
+     */
+    public boolean addURLToDownload(TokenedUrlGetter tug, RipUrlId ripUrlId, Path directory, String filename, String referrer, Map<String,String> cookies, Boolean getFileExtFromMIME) {
+        // Only download one file if this is a test.
+        if (isThisATest() && (itemsCompleted.size() > 0 || itemsErrored.size() > 0)) {
+            stop();
+            itemsPending.clear();
+            return false;
+        }
+        if (!allowDuplicates()
+                && ( itemsPending.contains(ripUrlId)
+                || itemsCompleted.containsKey(ripUrlId)
+                || itemsErrored.containsKey(ripUrlId) )) {
+            // Item is already downloaded/downloading, skip it.
+            // TODO print path if in itemsCompleted or itemsErrored
+            logger.info("[!] Skipping " + ripUrlId + " -- already attempted: " + Utils.removeCWD(directory));
+            return false;
+        }
+
+        if (Utils.getConfigBoolean("urls_only.save", false)) {
+            // Output URL to file
+            Path urlFile = Paths.get(this.workingDir + "/urls.txt");
+            URL url = null;
+            try {
+                url = tug.getTokenedUrl();
+            } catch (IOException | URISyntaxException e) {
+                logger.error("Unable to get URL for {}", ripUrlId, e);
+                itemsErrored.put(ripUrlId, e.getMessage());
+                return false;
+            }
+            if (AbstractRipper.shouldIgnoreExtension(url)) {
+                sendUpdate(STATUS.DOWNLOAD_SKIP, "Skipping " + url.toExternalForm() + " - ignored extension");
+                return false;
+            }
+            String text = url.toExternalForm() + System.lineSeparator();
+            try {
+                Files.write(urlFile, text.getBytes(StandardCharsets.UTF_8), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+                itemsCompleted.put(ripUrlId, urlFile);
+            } catch (IOException e) {
+                logger.error("Error while writing to " + urlFile, e);
+                return false;
+            }
+            return true;
+        }
+        else {
+            itemsPending.add(ripUrlId);
+            DownloadFileThread dft = new DownloadFileThread(tug, ripUrlId, directory, filename, this, getFileExtFromMIME);
+            if (referrer != null) {
+                dft.setReferrer(referrer);
+            }
+            if (cookies != null) {
+                dft.setCookies(cookies);
+            }
+            threadPool.addThread(dft);
+        }
+
+        return true;
+    }
+
 
 
     /**
@@ -591,7 +654,7 @@ public abstract class AbstractRipper
             return;
         }
 
-        if (!completed && itemsPending.isEmpty()) {
+        if (!completed && itemsPending.isEmpty()) { // TODO add itemsActive for current transfers
             completed = true;
             logger.info("   Rip completed!");
